requirements.txt
Django==4.2
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.1
sentence-transformers==2.2.2
feedparser==6.0.10
line-bot-sdk==1.20.0
gunicorn==21.2.0
